fixed bugs, added documentation, split files
